+++ pages/components/email-app/index.ts
@@ -1,0 +1,1 @@
+export { EmailApp } from "./email-app";
