+++ pages/components/open-modal/index.ts
@@ -1,0 +1,1 @@
+export { OpenModal } from "./open-modal";
