--- pages/dashboard/components/index.ts
@@ -1,3 +1,0 @@
-export { EmailApp } from "./email-app";
-export { OpenModal } from "./open-modal";
-export { StyledPopup, Modal, ButtonWrapper, CancelButton } from "./modal.style";
