+++ pages/index.jsx
@@ -1,9 +1,10 @@
 import styled from "styled-components";
 import { Routes } from "@config/routes";
-import * as M from "./dashboard/components";
 import Popup from "reactjs-popup";
-import { EmailApp } from "./dashboard/components/email-app";
-import { OpenModal } from "./dashboard/components/open-modal";
+import { Button } from "@features/ui";
+import { EmailApp } from "./components/email-app";
+import { OpenModal } from "./components/open-modal";
+import { color, space, textFont } from "@styles/theme";
 
 const Header = styled.header`
   width: 100%;
@@ -32,6 +33,85 @@ const ContactButton = styled.button`
   }
 `;
 
+const StyledPopup = styled(Popup)`
+  &-overlay {
+    display: none;
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background: rgba(52, 64, 84, 0.6);
+    backdrop-filter: blur(8px);
+
+    /* animation */
+
+    @keyframes anvil {
+      0% {
+        transform: scale(1) translateY(0px);
+        opacity: 0;
+        box-shadow: 0 0 0 rgba(241, 241, 241, 0);
+      }
+      1% {
+        transform: scale(0.96) translateY(10px);
+        opacity: 0;
+        box-shadow: 0 0 0 rgba(241, 241, 241, 0);
+      }
+      100% {
+        transform: scale(1) translateY(0px);
+        opacity: 1;
+        box-shadow: 0 0 500px rgba(241, 241, 241, 0);
+      }
+    }
+    .popup-content {
+      -webkit-animation: anvil 0.3s cubic-bezier(0.38, 0.1, 0.36, 0.9) forwards;
+    }
+  }
+`;
+
+const Modal = styled.div`
+  width: 25em;
+  height: 18em;
+  padding: 0;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+  align-items: center;
+  background: white;
+  border-radius: 12px;
+  box-shadow: 0px 20px 24px -4px rgba(16, 24, 40, 0.1),
+    0px 8px 8px -4px rgba(16, 24, 40, 0.04);
+  z-index: 1;
+
+  @media (max-width: 26.563em) {
+    width: 21.4375em;
+  }
+`;
+const ButtonWrapper = styled.div`
+  display: flex;
+  flex-direction: row;
+  padding-top: ${space(8)};
+`;
+
+const CancelButton = styled(Button)`
+  color: ${color("gray", 700)};
+  padding: 0.625rem 3.66rem;
+  border: 1px solid ${color("gray", 300)};
+  border-radius: 0.5rem;
+  background: white;
+  margin-right: 0.375rem;
+  ${textFont("md", "medium")}
+
+  &:hover {
+    background: ${color("gray", 50)};
+  }
+
+  @media (max-width: 26.563em) {
+    padding: 0.625rem 2.766rem;
+  }
+`;
+
 const IssuesPage = () => {
   return (
     <div>
@@ -43,7 +123,7 @@ const IssuesPage = () => {
         <a href={Routes.projects}>Dashboard</a>
       </Header>
 
-      <M.StyledPopup
+      <StyledPopup
         trigger={
           <ContactButton data-cy="contact-button">
             {/* eslint-disable-next-line @next/next/no-img-element */}
@@ -55,7 +135,7 @@ const IssuesPage = () => {
       >
         {(close) => (
           <div>
-            <M.Modal>
+            <Modal>
               <OpenModal
                 className="popup-content"
                 src="/icons/modal.svg"
@@ -65,8 +145,8 @@ const IssuesPage = () => {
                 answer within 24 hours."
               />
 
-              <M.ButtonWrapper>
-                <M.CancelButton
+              <ButtonWrapper>
+                <CancelButton
                   onClick={() => {
                     console.log("modal closed ");
                     close();
@@ -74,7 +154,7 @@ const IssuesPage = () => {
                   data-cy="cancel-button"
                 >
                   Cancel
-                </M.CancelButton>
+                </CancelButton>
                 <div>
                   <Popup
                     trigger={
@@ -85,11 +165,11 @@ const IssuesPage = () => {
                     }
                   ></Popup>
                 </div>
-              </M.ButtonWrapper>
-            </M.Modal>
+              </ButtonWrapper>
+            </Modal>
           </div>
         )}
-      </M.StyledPopup>
+      </StyledPopup>
     </div>
   );
 };
